# Conflicts:
#	.gitignore

+++ samples/autonomous-setup/bin/stop_agent.sh
@@ -5,14 +5,48 @@ DIR_PREFIX=$(dirname "$0")
 
 initLocations "$DIR_PREFIX"
 
-if [ -f "$AGENT_PID" ] ; then
-  PID=$(cat "$AGENT_PID")
-  kill -3 $PID
-fi
+sighupAgent() {
+  if [ -f "$AGENT_PID" ]; then
+    PID=$(cat "$AGENT_PID")
 
-sleep 2
+    PROC_CNT=$(ps aux | grep $PID)
+
+    if [ $PROC_CNT != 0 ]; then
+      kill -3 $PID
+      sleep 2
+    else
+      echo "Agent isn't currently running"
+    fi
+  fi
+}
+
+sighupAgentCygwin() {
+  if [ -f "$AGENT_PID" ]; then
+    PID=$(cat "$AGENT_PID")
+    CYGWIN_PID=$(ps aux | grep $PID | awk '{print $1}')
+
+    if [ "$CYGWIN_PID" != "" ]; then
+      kill -3 $CYGWIN_PID
+      sleep 2
+    else
+      echo "Agent isn't currently running"
+    fi
+  fi
+}
+
+case "$(uname)" in
+CYGWIN*)
+  sighupAgentCygwin
+  ;;
+
+*)
+  sighupAgent
+  ;;
+esac
+
+
 
 NUM_REMAINING=$(ps aux | grep vault | grep agent | wc -l)
-if [ $NUM_REMAINING -gt 0 ] ; then
-   echo: "Warning: Vault agent didn't exit gracefully. You need to shut it down manually"
+if [ $NUM_REMAINING -gt 0 ]; then
+  echo: "Warning: Vault agent didn't exit gracefully. You need to shut it down manually"
 fi
